backend/requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.2
django-cors-headers==3.13.0
django-import-export==2.5.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.3.0
docxcompose==1.3.5
docxtpl==0.16.4
et-xmlfile==1.1.0
jdcal==1.4.1
Jinja2==3.1.2
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.1
numpy==1.22.2
odfpy==1.4.1
openpyxl==3.0.4
pandas==1.4.1
pdfkit==1.0.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
